meteo-0.9.0.tar.gz

+++ src/meteopoll.cc
@@ -24,10 +24,11 @@
 #include <sys/wait.h>
 #endif
 #include <Station.h>
+#include <StationFactory.h>
 #include <Configuration.h>
 #include <MeteoException.h>
 #include <Datasink.h>
-#include <Pidfile.h>
+#include <Daemon.h>
 #include <mdebug.h>
 #include <errno.h>
 #ifdef HAVE_STRING_H
@@ -42,38 +43,57 @@
 #ifdef HAVE_SIGNAL_H
 #include <signal.h>
 #endif
+#include <sys/resource.h>
 
 #define	ALARMTIMEOUT	30
 
 pid_t	clpid;
+bool	allowfork = true;;
 
 // signal handler to kill child when parent is sent a signal
 void	kill_child(int cause) {
 	switch (cause) {
 	case SIGTERM:
 	case SIGINT:
-		kill(clpid, cause);
+		if (allowfork)
+			kill(clpid, cause);
 		mdebug(LOG_INFO, MDEBUG_LOG, 0, "caught signal, exiting");
 		exit(EXIT_SUCCESS);
 		break;
 	case SIGHUP:
-		mdebug(LOG_INFO, MDEBUG_LOG, 0, "caught SIGHUP, killing child");
-		kill(clpid, SIGTERM);
+		mdebug(LOG_INFO, MDEBUG_LOG, 0,
+			"caught SIGHUP, killing child");
+		if (allowfork)
+			kill(clpid, SIGTERM);
 		break;
 	}
 	return;
 }
 
-static void	loop(const meteo::Configuration& conf, const std::string& name) {
+static void	loop(const std::string& station) {
+	// limit memory to 16MB so this process cannot monopolize the system
+#define	memmax	0x1 << 25
+	struct rlimit	l;
+	getrlimit(RLIMIT_STACK, &l);	// stack segment
+	if (l.rlim_max > memmax)
+		l.rlim_max = memmax;
+	setrlimit(RLIMIT_STACK, &l);
+
+	getrlimit(RLIMIT_DATA, &l);	// data segment
+	if (l.rlim_max > memmax)
+		l.rlim_max = memmax;
+	setrlimit(RLIMIT_DATA, &l);
+
 	// create a new station object (this always returns something
 	// connected)
 	mdebug(LOG_DEBUG, MDEBUG_LOG, 0, "starting new station '%s'",
-		name.c_str());
-	meteo::Station	*s = meteo::StationFactory(conf).newStation(name);
-	mdebug(LOG_DEBUG, MDEBUG_LOG, 0, "station %s ready", name.c_str());
+		station.c_str());
+	meteo::Station	*s = meteo::StationFactory().newStation(station);
+	mdebug(LOG_DEBUG, MDEBUG_LOG, 0, "station %s ready", station.c_str());
 
 	// create a data sink
-	meteo::Datasink	*ds = meteo::DatasinkFactory().newDatasink(name);
+	meteo::Datasink	*ds = meteo::DatasinkFactory().newDatasink(station);
+	mdebug(LOG_DEBUG, MDEBUG_LOG, 0, "datasink %s ready", station.c_str());
 
 	// loop:
 	meteo::Timeval	looptime; looptime.now();
@@ -105,10 +125,12 @@ static void	loop(const meteo::Configuration& conf, const std::string& name) {
 			// reset the datarecord
 			looptime.now();
 			if (minute != looptime.getMinute()) {
-				// send data to the database
-				std::cout << "sending update to db: "
-					<< s->updatequery() << std::endl;
-				ds->receive(s->updatequery());
+				// get the lists of queries
+				std::list<std::string>	queries
+					= s->updatequery(looptime.getTimekey());
+
+				// send each query to the database
+				ds->receive(queries);
 
 				// reset and restart
 				s->reset();
@@ -123,7 +145,7 @@ static void	usage(void) {
 "usage: meteopoll [ -dFVh? ] [ -l logurl ] [ -f conffile ] [ -b prefs ] \\\n"
 "       [ -p pidfile ] -s stationname\n"
 "   -d                increase debug level\n"
-"   -F                don't fork (used for debuggin)\n"
+"   -F                don't fork (used for debugging)\n"
 "   -V                print version and exit\n"
 "   -h, -?            print this help screen and exit\n"
 "   -f conffile       use conffile, see meteo.xml(5)\n"
@@ -136,10 +158,9 @@ static void	usage(void) {
 int	main(int argc, char *argv[]) {
 	std::string	conffile(METEOCONFFILE);
 	std::string	logurl("file:///-");	// logging to stderr
-	std::string	name("undefined");
-	std::vector<std::string>	preferences;
-	std::string	pidfilename;
-	bool		allowfork = true;;
+	std::string	station;
+	meteo::stringlist	preferences;
+	std::string	pidfileprefix;
 
 	// parse command line
 	int	c;
@@ -155,13 +176,13 @@ int	main(int argc, char *argv[]) {
 			conffile = std::string(optarg);
 			break;
 		case 's':
-			name = std::string(optarg);
+			station = std::string(optarg);
 			break;
 		case 'b':
 			preferences.push_back(std::string(optarg));
 			break;
 		case 'p':
-			pidfilename = optarg;
+			pidfileprefix = std::string(optarg);
 			break;
 		case 'V':
 			fprintver(stdout);
@@ -186,39 +207,18 @@ int	main(int argc, char *argv[]) {
 		logurl.c_str());
 
 	// it is an error if station name or conf
-	if ("undefined" == name) {
+	if (station.empty()) {
 		fprintf(stderr, "%s: station name not set\n", argv[0]);
 		mdebug(LOG_CRIT, MDEBUG_LOG, 0, "station name not set");
 		exit(EXIT_FAILURE);
 	}
 
+	// pid file
+	if (pidfileprefix.empty())
+		pidfileprefix = "/var/run/meteopoll-";
+
 	// fork if necessary
-	if (allowfork) {
-		// we are allowed to fork, so do it
-		pid_t	pid = fork();
-		if (pid <  0) {
-			mdebug(LOG_DEBUG, MDEBUG_LOG, MDEBUG_ERRNO,
-				"cannot fork");
-			exit(EXIT_FAILURE);
-		}
-		if (pid > 0) { exit(EXIT_SUCCESS); }
-
-		// do what is necessary to initialize a daemon
-		setsid();
-		chdir("/");
-		umask(022);
-	}
-
-	// XXX install a signal handler that will kill all child processes
-	//     if we receive a deadly signal
-
-	// write a pid file
-	if ("" == pidfilename) {
-		pidfilename = "/var/run/meteopoll-" + name + ".pid";
-	}
-	if (allowfork) {
-		meteo::Pidfile	pf(pidfilename);
-	}
+	meteo::Daemon	d(pidfileprefix, station, !allowfork);
 
 	// open the configuration object
 	meteo::Configuration	conf(conffile);
@@ -229,7 +229,7 @@ int	main(int argc, char *argv[]) {
 	if (0 == preferences.size()) {
 		preferences.push_back("debug");
 	}
-	meteo::DatasinkFactory	dsf(conf, preferences);
+	meteo::DatasinkFactory	dsf(preferences);
 
 	// install signal handlers
 	signal(SIGTERM, kill_child);
@@ -239,18 +239,28 @@ int	main(int argc, char *argv[]) {
 	// loop: fork a process and wait for it to die, exit if the child
 	// exits naturally
 	int	status;
+	bool	giveup = false;
 	do {
 		mdebug(LOG_DEBUG, MDEBUG_LOG, 0, "start loop");
-		clpid = fork();
+
+		// fork only if allowed to do so
+		if (allowfork)
+			clpid = fork();
+		else
+			clpid = 0;
+
 		// if the fork is not possible, we have some serious problem
 		if (clpid < 0) {
 			mdebug(LOG_CRIT, MDEBUG_LOG, MDEBUG_ERRNO,
 				"fork failed: %s", strerror(errno));
 			exit(EXIT_FAILURE);
 		}
+
 		// the child simply starts the loop
 		if (clpid == 0) {
-			mdebug(LOG_DEBUG, MDEBUG_LOG, 0, "starting new child");
+			if (allowfork)
+				mdebug(LOG_DEBUG, MDEBUG_LOG, 0,
+					"starting new child");
 			// reset signal dispositions so that child can easily
 			// be killed
 			signal(SIGHUP, SIG_DFL);
@@ -259,21 +269,66 @@ int	main(int argc, char *argv[]) {
 
 			// start looping
 			try {
-				loop(conf, name);
+				mdebug(LOG_DEBUG, MDEBUG_LOG, 0,
+					"starting loop(%s)", station.c_str());
+				loop(station);
+				mdebug(LOG_DEBUG, MDEBUG_LOG, 0,
+					"return from loop");
 			} catch (meteo::MeteoException& e) {
 				mdebug(LOG_ERR, MDEBUG_LOG, 0,
 					"MeteoException(%s, %s)", 
 					e.getReason().c_str(),
 					e.getAddinfo().c_str());
+				// when we exit at this point, the parent
+				// process will fork a new child (if there is a
+				// a parent, that is), and it will know that we
+				// had a problem
 				exit(EXIT_FAILURE);
 			}
+			// exiting at this point means that for some reason loop
+			// ended without error. This should not really happen,
+			// but to tell the parent that something different happend
+			// we exit with EXIT_SUCCESS
 			exit(EXIT_SUCCESS);
-		}
+		} // end of child code
+
+		// here we wait for the child to terminate
 		if (waitpid(clpid, &status, 0) < 0) {
-			mdebug(LOG_CRIT, MDEBUG_LOG, MDEBUG_ERRNO,
-				"child process died: ");
+			mdebug(LOG_ERR, MDEBUG_LOG, 0, "error during wait: %s",
+				strerror(errno));
+		} else {
+			// here we should be a bit more verbose, so that it
+			// becomes less difficult to find out what happened
+			// to the child
+			if (WIFEXITED(status)) {
+				mdebug(LOG_DEBUG, MDEBUG_LOG, 0,
+					"child %d exited with status %d",
+					clpid, WEXITSTATUS(status));
+				giveup = (WEXITSTATUS(status) == 0);
+			}
+			if (WIFSIGNALED(status)) {
+				mdebug(LOG_DEBUG, MDEBUG_LOG, 0,
+					"child %d killed by signal %d",
+					clpid, WTERMSIG(status));
+			}
+			if (WIFSTOPPED(status)) {
+				// this should definitely not happen, lets kill
+				// the process and start a new one
+				mdebug(LOG_DEBUG, MDEBUG_LOG, 0,
+					"child %d stopped by signal %d",
+					clpid, WSTOPSIG(status));
+				kill(clpid, SIGKILL);
+			}
+			mdebug(LOG_CRIT, MDEBUG_LOG, 0,
+				"child process died: %d", WEXITSTATUS(status));
 		}
-	} while (!(WIFEXITED(status) && (WEXITSTATUS(status) == 0)));
+		// if there is a problem with the station, we should wait a
+		// few seconds so the station can recover from the problem.
+		// This also prevents meteopoll from looping. 
+		mdebug(LOG_DEBUG, MDEBUG_LOG, 0, "sleeping for 3 seconds to "
+			"give station time to recover from error condition");
+		sleep(3);
+	} while (!giveup);
 
 	// child exited ok, so we do the same
 	exit(EXIT_SUCCESS);
